refactor(engine): move the O1 funding-coverage guard to orders/coverage.ts
Pure extraction of ingest.ts:564-738 into a dedicated coverage module, per audit finding 33; the barrel export, README rows, and the three test files' imports are re-pointed to the new module, and behavior is unchanged.
